+++ tests/test_visitors/test_tokenize/test_primitives/test_numbers/test_underscored_numbers.py
@@ -11,11 +11,27 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
 @pytest.mark.parametrize(
     'primitive',
     [
-        '333_555',
+        '1_00',
+        '10_00',
+        '100_00',
+        '-1_0',
+        '-10_0',
+        '-100_0',
+        '1000_000',
+        '-1000_000',
+        '1.0_0',
         '3_3.3',
-        '+1_000_000',
-        '-5_000',
-        '-1_000.0',
+        '1_0_0',
+        '1_0_0_1',
+        '1_0_0.1',
+        '0_0.00_1',
+        '0.000_1',
+        '0.00_01',
+        '100_000.000_01',
+        '100_00.000_001',
+        '1000_000.000_001',
+        '10_000_000.000000001',
+        '10000000.000_000_001',
     ],
 )
 def test_underscored_number(
@@ -27,7 +43,7 @@ def test_underscored_number(
     primitive,
     mode,
 ):
-    """Ensures that underscored numbers raise a warning."""
+    """Ensures that wrongly underscored numbers raise a warning."""
     file_tokens = parse_tokens(mode(primitives_usages.format(primitive)))
 
     visitor = WrongNumberTokenVisitor(default_options, file_tokens=file_tokens)
@@ -44,7 +60,33 @@ def test_underscored_number(
         '1000.0',
         '-333555',
         '-333555.5',
-        '"10_00"',
+        '1_000',
+        '1_000.0',
+        '+1_000',
+        '+1_000.0',
+        '-1_000',
+        '-1_000.0',
+        '10_000',
+        '100_000',
+        '100_000_000',
+        '0.0_005',
+        '100_000.000_001',
+        '10_000_000.000_000_001',
+        '1_000_000.000_000_001',
+        '100_000.000_000_001',
+        '10_000.000_000_001',
+        '1_000.000_000_001',
+        '100.000_000_001',
+        '10.000_000_001',
+        '1.000_000_001',
+        '10_000_000.00_000_001',
+        '10_000_000.0_000_001',
+        '10_000_000.000_001',
+        '10_000_000.00_001',
+        '10_000_000.0_001',
+        '10_000_000.001',
+        '10_000_000.01',
+        '10_000_000.1',
     ],
 )
 def test_correct_number(