+++ CHANGELOG.md
@@ -75,6 +75,7 @@ Semantic versioning in our case means:
 - Allows walrus operator (`:=`) in comprehesions #3121
 - Does not count `self`, `cls`, and `mcs` as arguments
   for `WPS211` complexity check anymore, #2394
+- `WPS303`: Underscore (_) now only allowed with 3 digits after it, #3120
 
 ### Bugfixes